Change to display the edit button and delete button only for the same author's article as the logged-in user

+++ resources/views/article/index.blade.php
@@ -15,12 +15,15 @@
                         <p class="lead">{{ $article->content }}</p>
                         <p class="text-right"><a href="{{ route('article.show', $article) }}" class="btn btn-default" role="button">More</a></p>
 
-                        <a href="{{ route('article.edit', $article) }}" type="button" class="btn btn-primary">Edit</a>
-                        {{ Form::model($article, ['route' => ['article.destroy', $article->id], 'method' => 'DELETE', 'class' => 'form-inline btn']) }}
-                        <div class="form-group">
-                            {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
-                        </div>
-                        {{ Form::close() }}
+                        @if($article->user->is_same_auth_user)
+                            <a href="{{ route('article.edit', $article) }}" type="button"
+                               class="btn btn-primary">Edit</a>
+                            {{ Form::model($article, ['route' => ['article.destroy', $article->id], 'method' => 'DELETE', 'class' => 'form-inline btn']) }}
+                            <div class="form-group">
+                                {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
+                            </div>
+                            {{ Form::close() }}
+                        @endif
                     </div>
                 </div>
             </div>
